Lab 4: Interact collider fix 1
Discovered that a game object could not have a collider that is and is not a trigger, even on a child of an object. Doing so broke jumping and other connected functions.

First fix included eliminating the interact sphere and adding/altering scripts to fix pickup functionality.

Second fix will focus on fixing functionality on other interactable items to work without interact sphere.